Misc code updates
