+++ app/src/AppRouter.tsx
@@ -44,7 +44,14 @@ const AppRouter: React.FC = (props: any) => {
         layout={AuthLayout}
         title={getTitle('Request submitted')}
       />
-      <AppRoute protected path="/admin" component={AdminRouter} layout={AuthLayout} title={getTitle('Admin')} />
+      <AppRoute
+        protected
+        path="/admin"
+        component={AdminRouter}
+        layout={AuthLayout}
+        title={getTitle('Admin')}
+        validRoles={['abc']}
+      />
       <AppRoute title="*" path="*" component={() => <Redirect to="/page-not-found" />} />
     </Switch>
   );
