+++ app/src/utils/PrivateRoute.tsx
@@ -1,6 +1,6 @@
-import React from 'react';
-import { Route, RouteProps, Redirect } from 'react-router-dom';
 import { AuthStateContext } from 'contexts/authStateContext';
+import React from 'react';
+import { Redirect, Route, RouteProps } from 'react-router-dom';
 
 interface IPrivateRouteProps extends RouteProps {
   validRoles?: string[];
